% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmea.R
\name{prep_data}
\alias{prep_data}
\title{Data prep for a graphical depiction of risk based on an FMEA.}
\usage{
prep_data(filename)
}
\arguments{
\item{filename}{FMEA filename as an imported data frame.}
}
\value{
Returns a modified data frame to be used in the graphical depiction of risk.
}
\description{
This function uses an imported data set and preps it for graphical analysis. Specifically, it adds a random jitter to severity, occurence and detection based on a standard normal variable (mean = 0 and sd = 0.25)
}
\note{
Geom_jitter is by-passed with this function to allow ggrepel to work more efficiently.
}
\examples{
df<-prep_data(fmea_worksheet_name)
}
