% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmea.R
\name{risk_controls}
\alias{risk_controls}
\title{Potting a scatterplot of risk management controls.}
\usage{
risk_controls(filename)
}
\arguments{
\item{filename}{Filename as an imported data frame (usually from an FMEA study). Must possess the fields named item_no, process_step, severity,occurrence, and detection.}
}
\value{
Returns a modified data frame to be used in the graphical depiction of risk controls.
}
\description{
This function uses an imported data set and creates a risk control graph based on detection and occurrence.
}
\note{
Geom_jitter is not needed with this function to allow ggrepel to work more efficiently with labels. Specifically, a random jitter is added to severity, occurence and detection based on a standard normal variable (mean = 0 and sd = 0.25)
}
\examples{
risk_controls(fmea_worksheet_name)
}
