% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmea.R
\name{risk_disutility}
\alias{risk_disutility}
\title{Plotting a disutility or hazard graph of risk.}
\usage{
risk_disutility(filename)
}
\arguments{
\item{filename}{Filename as an imported data frame (usually from an FMEA study). Must possess the fields named item_no, process_step, severity,occurrence, and detection.}
}
\value{
Returns a modified data frame to be used in the graphical depiction of risk.
}
\description{
This function uses an imported data set and creates a disutility graph.
}
\note{
Geom_jitter is not needed with this function to allow ggrepel to work more efficiently with labels. Specifically, a random jitter is added to severity, occurence and detection based on a standard normal variable (mean = 0 and sd = 0.25)
}
\examples{
\dontrun{

# Use the working file with a standard format to create a hazard or disutility graphic.

library(ggplot2)
library(dplyr)
library(ggrepel)

# If importing an Excel worksheet, the following code chunk is to be used (based on your working directory).

library(readxl)
fmea_working_file <- read_excel("~/Data/Statistical Projects/R Projects/FMEA/inst/extdata/fmea_working_file.xls")

risk_disutility(fmea_working_file)

}
}
